Rename infer-schema -> introspector
