+++ docs/modules/ROOT/pages/troubleshooting/index.adoc
@@ -22,13 +22,13 @@ Alternatively, if you are debugging a particular functionality, you can specify 
 3. `@neo4j/graphql:graphql` - Logs the GraphQL query and variables
 4. `@neo4j/graphql:execute` - Logs the Cypher and Cypher paramaters before execution, and summary of execution
 
-=== For @neo4j/infer-schema
+=== For @neo4j/introspector
 
-`@neo4j/infer-schema` has its own debug logging namespace and you can turn on logging for it with:
+`@neo4j/introspector` has its own debug logging namespace and you can turn on logging for it with:
 
 [source, bash, indent=0]
 ----
-DEBUG=@neo4j/infer-schema node src/index.js
+DEBUG=@neo4j/introspector node src/index.js
 ----
 
 [[troubleshooting-query-tuning]]
