+++ docs/modules/ROOT/pages/type-definitions/introspector.adoc
@@ -1,7 +1,7 @@
-[[type-definitions-infer-schema]]
-= Infer schema from an existing database
+[[type-definitions-introspector]]
+= Introspect schema from an existing Neo4j database
 
-Neo4j provides a tool that enables you to, with very little effort, generate GraphQL type definitions from an existing database. This is provided from a separate npm package `@neo4j/infer-schema`.
+Neo4j provides a tool that enables you to, with very little effort, generate GraphQL type definitions from an existing database. This is provided from a separate npm package `@neo4j/introspector`.
 
 This is usually a one-time-thing and should be considered a starting point for a GraphQL schema.
 
@@ -35,7 +35,7 @@ You can then serve this file with your GraphQL server.
 
 [source, javascript, indent=0]
 ----
-const { toGraphQLTypeDefs } = require("@neo4j/infer-schema")
+const { toGraphQLTypeDefs } = require("@neo4j/introspector")
 const neo4j = require("neo4j-driver");
 const fs = require('fs');
 
@@ -65,7 +65,7 @@ Here the type definitions are never persisted to disk.
 [source, javascript, indent=0]
 ----
 const { Neo4jGraphQL } = require("@neo4j/graphql");
-const { toGraphQLTypeDefs } = require("@neo4j/infer-schema")
+const { toGraphQLTypeDefs } = require("@neo4j/introspector")
 const neo4j = require("neo4j-driver");
 
 const driver = neo4j.driver(
