+++ jest.config.base.js
@@ -14,7 +14,7 @@ module.exports = {
     moduleFileExtensions: ["js", "json", "jsx", "ts", "tsx", "node", "md"],
     moduleNameMapper: {
         "@neo4j/graphql/dist/types": "<rootDir>/packages/graphql/src/types",
-        "@neo4j/infer-schema(.*)$": "<rootDir>/packages/infer-schema/src/$1",
+        "@neo4j/introspector(.*)$": "<rootDir>/packages/introspector/src/$1",
         "@neo4j/graphql-ogm(.*)$": "<rootDir>/packages/ogm/src/$1",
         "@neo4j/graphql(.*)$": "<rootDir>/packages/graphql/src/$1",
     },
