--- packages/infer-schema/README.md
@@ -1,1 +1,0 @@
-## Infer Schema
