+++ packages/introspector/.npmignore
