+++ packages/introspector/.npmrc
