+++ packages/introspector/src/classes/Node.ts
