+++ packages/introspector/src/classes/Property.ts
