+++ packages/introspector/src/classes/Relationship.ts
