+++ packages/introspector/src/to-internal-struct.ts
