+++ packages/introspector/src/transforms/neo4j-graphql/directives/RelationshipProperties.ts
