+++ packages/introspector/src/transforms/neo4j-graphql/utils/infer-relationship-props-name.ts
