+++ packages/introspector/src/utils/pascal-case.test.ts
