+++ packages/introspector/src/utils/pascal-case.ts
