+++ packages/introspector/tsconfig.json
