+++ tsconfig.json
@@ -3,7 +3,7 @@
     "references": [
         { "path": "./packages/graphql/src/tsconfig.json" },
         { "path": "./packages/ogm/src/tsconfig.json" },
-        { "path": "./packages/infer-schema/src/tsconfig.json" },
+        { "path": "./packages/introspector/src/tsconfig.json" },
         { "path": "./examples/neo-push/client/src/tsconfig.json" },
         { "path": "./examples/neo-push/server/src/tsconfig.json" }
     ]
